Struct rustic_hal::resource::OneOrMany

pub struct OneOrMany<T> { /* fields omitted */ }

A Simple wrapper around a vector to allow custom serialization when only 1 element is contained.

example

In the example below, the vector serializes to json as an object if it contains only one value, but as an array if more than one.

use rustic_hal::resource::OneOrMany;
use rustic_hal::HalLink;

use serde_json::to_string;
let mut v = OneOrMany::new();
v.push(&HalLink::new("http://test.com"));

assert_eq!(to_string(&v).unwrap(), r#"{"href":"http://test.com"}"#);

v.push(&HalLink::new("http://test2.com"));

assert_eq!(to_string(&v).unwrap(), r#"[{"href":"http://test.com"},{"href":"http://test2.com"}]"#);Run

Methods

impl<T> OneOrMany<T> where T: Clone

fn new() -> OneOrMany<T>

create a new empty object

fn force_many(&mut self) -> &mut Self

Force to be serialized as array, even if only one element

fn len(&self) -> usize

retrieve the length of the wrapped vector

fn is_empty(&self) -> bool

fn single(&self) -> Option<&T>

Retrieves a single element if possible.

fn many(&self) -> &Vec<Box<T>>

Returns an immutable reference to the contained links

fn push(&mut self, newval: &T)

Add an element to the wrapped vector.